pdf-ETL/pyproject.toml
[tool.poetry]
name = "pdf-etl"
version = "0.1.0"
description = ""
authors = ["Davi Alvarenga <davialvarenga@hotmail.com.br>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
camelot-py = "^0.11.0"
opencv-python = "^4.10.0.84"
matplotlib = "^3.9.2"
ghostscript = "^0.7"
pandas = "^2.2.2"
psycopg2 = "^2.9.9"
psycopg2-binary = "^2.9.9"
sqlalchemy = "^2.0.32"
unicode = "^2.9"
unidecode = "^1.3.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
